+++ torch2trt/converters/relu.py
@@ -6,6 +6,7 @@ from torch2trt.module_test import add_module_test
 @tensorrt_converter('torch.relu_')
 @tensorrt_converter('torch.nn.functional.relu')
 @tensorrt_converter('torch.nn.functional.relu_')
+@tensorrt_converter('torch.Tensor.relu')
 def convert_functional_relu(ctx):
     ctx.method_args = (torch.nn.ReLU(),) + ctx.method_args
     convert_relu(ctx)
@@ -33,3 +34,16 @@ class FunctionalRelu(torch.nn.Module):
 @add_module_test(torch.float32, torch.device('cuda'), [(1, 3, 4, 5)])
 def test_functional_relu_basic():
     return FunctionalRelu()
+
+
+class TensorRelu(torch.nn.Module):
+    def __init__(self):
+        super(TensorRelu, self).__init__()
+
+    def forward(self, x):
+        return x.relu()
+
+
+@add_module_test(torch.float32, torch.device('cuda'), [(1, 3, 40, 20)])
+def test_tensor_relu():
+    return TensorRelu()