+++ CHANGELOG.md
@@ -10,3 +10,6 @@
 - Added GroupNorm plugin which internally uses PyTorch aten::group_norm
 - Replaced Tensor.ndim references with len(tensor.shape) to support older pytorch versions
 - Added reduced precision documentation page
+- Added converters for ``floordiv``, ``mod``, ``ne``, and ``torch.tensor`` operations
+- Extended ``relu`` converter to support ``Tensor.relu`` operation
+- Extended ``sigmoid`` converter to support ``Tensor.sigmoid`` operation